feat: create Factory
